requirements.txt
django==6.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
Pillow==10.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0

